[backend/양진영][W8-T6] GraphQL 예외처리(Validation) (#613)
* [backend/양진영][#598] refactor: 기존api 레이어명 변경(controller)

* [backend/양진영][#598] refactor: admin 비지니스 로직 통합

* [backend/양진영][#598] refactor: user 비지니스 로직 통합

* [backend/양진영][#598] refactor: book 비지니스 로직 통합

* [backend/양진영][#598] refactor: bookInfo, bookCategory 비지니스 로직 통합

* [backend/양진영][#598] refactor: rental 비지니스 로직 통합

* [backend/양진영][#598] refactor: reservation 비지니스 로직 통합

* [backend/양진영][#536] feat: 에러코드 상수화

* [backend/양진영][#536] feat: API별 일관된 에러응답을 위한 응답 인터페이스 로직 추가

* [backend/양진영][#536] feat: Apollo 서버 에러응답을 위한 로직 추가

* [backend/양진영][#536] feat: 에러응답 로직 수정 및 에러 코드 추가

- 에러코드 추가 및 status code, 메시지 맵핑
- 에러객체 호출 시, 메시지를 담을 수 있도록 수정
- 에러응답 모듈명 변경(responseError => ErrorResponse)

* [backend/양진영][#536] feat: 변경된 에러핸들러 비지니스 로직 반영(Service 레이어)

* [backend/양진영][#536] feat: 에러핸들러 변경부분 컨트롤러 레이어 적용

* [backend/양진영][#536] feat: 기능 테스트 및 수정사항 반영

- Validation 에러인 경우, originMessage가 null임을 명시
- DB 참조관계 Validation 에러 미반영 부분 수정

* [backend/양진영][#536] fix: 오타 수정

- 유효하지 않다는 의미의 영단어(invalid)를 invalied로 입력된 부분 오타 수정

* [backend/양진영][#536] feat: 에러핸들러 Auth Validation 에러코드 추가

* [backend/양진영][#536] feat: 에러 호출 모듈(throwError) 제거

- 커스텀 에러객체 그대로 export하도록 수정

* [backend/양진영][#536] feat: 에러호출 모듈 제거 컨트롤러 레이어 반영

* [backend/양진영][#536] feat: 에러객체 호출 모듈 제거 서비스 레이어 반영

* [backend/양진영][#536] feat: 에러핸들러 변경부분 auth 로직 반영

* [backend/양진영][#422] feat: gql resolver 레이어 예외처리 추가

* [backend/양진영][#422] feat: Repository Admin, Book 로직 Validation 추가

* [backend/양진영][#422] feat: Repository bookCategory, bookInfo, rental, reservation, user Validation 추가

* [backend/양진영][#422] feat: 비지니스 로직(서비스 레이어) Validation 추가

* [backend/양진영][#536] feat: REST API 에러 응답 미들웨어 구현

* [backend/양진영][#536] feat: errorHanling 모듈 이름 변경(errorHandler)

* [backend/양진영][#536] feat: errorResponse 별도 레이어 추가에 따른 gql 로직 반영

* [backend/양진영][#422] feat: errorHandler 모듈명 변경 로직 반영

* [backend/양진영][#536] feat: 로그인 요청 실패 에러코드 변경

- 어떤 원인으로 실패했는지 노출되지 않도록 수정(보안관련)

* [backend/양진영][#536] feat: 유틸레이어에서 바로 Custom Error객체를 호출하도록 호출방식 변경

* [backend/양진영][#422] feat: 이전 브랜치 PR 리뷰 수정사항 반영

* [backend/양진영][#536] feat: 에러컨벤션 미반영 로직 적용(util)

+++ backend/jinyeong/repository/book.js
@@ -1,4 +1,8 @@
 const { Book } = require('../db');
+const { util, constants } = require('../common');
+
+const { errorHandler } = util;
+const { ERROR_CODE } = constants;
 
 const getListAll = async () => {
   const bookList = await Book.findAll();
@@ -10,6 +14,10 @@ const getListAll = async () => {
 
 // Book SELECT ALL IN ids
 const getAllByIds = async (ids) => {
+  if (!Array.isArray(ids)) {
+    throw new errorHandler.CustomError(ERROR_CODE.INTERNAL_SERVER_ERROR);
+  }
+
   const bookList = await Book.findAll({ where: { id: ids } });
 
   return bookList.map((book) => {
@@ -18,20 +26,35 @@ const getAllByIds = async (ids) => {
 };
 
 const getOneById = async (id) => {
+  if (!id) {
+    throw new errorHandler.CustomError(ERROR_CODE.INTERNAL_SERVER_ERROR);
+  }
+
   const book = await Book.findOne({ where: { id } });
   return book?.dataValues;
 };
 
 const createBook = async (inputData) => {
+  if (typeof inputData !== 'object') {
+    throw new errorHandler.CustomError(ERROR_CODE.INTERNAL_SERVER_ERROR);
+  }
+
   await Book.create(inputData);
 };
 
 const updateBook = async (id, inputData) => {
+  if (!id || typeof inputData !== 'object') {
+    throw new errorHandler.CustomError(ERROR_CODE.INTERNAL_SERVER_ERROR);
+  }
+
   await Book.update(inputData, { where: { id } });
 };
 
 const deleteBook = async (id) => {
-  // TODO: 누가, 언제 지웠는지 로그 쌓을 수 있도록 구현하기
+  if (!id) {
+    throw new errorHandler.CustomError(ERROR_CODE.INTERNAL_SERVER_ERROR);
+  }
+
   await Book.destroy({ where: { id } });
 };
