+++ backend/jinyeong/repository/bookCategory.js
@@ -1,4 +1,8 @@
 const { BookCategory } = require('../db');
+const { util, constants } = require('../common');
+
+const { errorHandler } = util;
+const { ERROR_CODE } = constants;
 
 const getListAll = async () => {
   const bookCategoryList = await BookCategory.findAll();
@@ -9,6 +13,10 @@ const getListAll = async () => {
 };
 
 const getAllByIds = async (ids) => {
+  if (!Array.isArray(ids)) {
+    throw new errorHandler.CustomError(ERROR_CODE.INTERNAL_SERVER_ERROR);
+  }
+
   const bookCategoryList = await BookCategory.findAll({ where: { id: ids } });
 
   return bookCategoryList.map((bookCategory) => {
@@ -17,16 +25,27 @@ const getAllByIds = async (ids) => {
 };
 
 const getOneById = async (id) => {
+  if (!id) {
+    throw new errorHandler.CustomError(ERROR_CODE.INTERNAL_SERVER_ERROR);
+  }
+
   const bookCategory = await BookCategory.findOne({ where: { id } });
   return bookCategory?.dataValues;
 };
 
 const createBookCategory = async (inputData) => {
+  if (typeof inputData !== 'object') {
+    throw new errorHandler.CustomError(ERROR_CODE.INTERNAL_SERVER_ERROR);
+  }
+
   await BookCategory.create(inputData);
 };
 
 const deleteBookCategory = async (id) => {
-  // TODO: 누가, 언제 지웠는지 로그 쌓을 수 있도록 구현하기
+  if (!id) {
+    throw new errorHandler.CustomError(ERROR_CODE.INTERNAL_SERVER_ERROR);
+  }
+
   await BookCategory.destroy({ where: { id } });
 };
